# internal/services/orderservice/internal/orders/features/getting_orders/v1/endpoints/get_orders_endpoint.go
package endpoints

import (
	"fmt"
	"net/http"

	"emperror.dev/errors"
	"github.com/DavidReque/go-food-delivery/internal/pkg/core/web/route"
	customErrors "github.com/DavidReque/go-food-delivery/internal/pkg/http/httperrors/customerrors"
	"github.com/DavidReque/go-food-delivery/internal/pkg/utils"
	"github.com/DavidReque/go-food-delivery/internal/services/orderservice/internal/orders/contracts/params"
	"github.com/DavidReque/go-food-delivery/internal/services/orderservice/internal/orders/features/getting_orders/v1/dtos"
	"github.com/DavidReque/go-food-delivery/internal/services/orderservice/internal/orders/features/getting_orders/v1/queries"

	"github.com/labstack/echo/v4"
	"github.com/mehdihadeli/go-mediatr"
)

type getOrdersEndpoint struct {
	params.OrderRouteParams
}

func NewGetOrdersEndpoint(params params.OrderRouteParams) route.Endpoint {
	return &getOrdersEndpoint{OrderRouteParams: params}
}

func (ep *getOrdersEndpoint) MapEndpoint() {
	ep.OrdersGroup.GET("", ep.handler())
}

// GetAllOrders
// @Tags Orders
// @Summary Get all orders
// @Description Get all orders with pagination and filters
// @Accept json
// @Produce json
// @Param page query int false "Page number" default(1) minimum(1)
// @Param size query int false "Page size" default(10) minimum(1) maximum(100)
// @Param orderBy query string false "Field to order by"
// @Param filters query string false "Applied filters"
// @Success 200 {object} dtos.GetOrdersResponseDto
// @Failure 400 {object} object
// @Router /api/v1/orders [get]
func (ep *getOrdersEndpoint) handler() echo.HandlerFunc {
	return func(c echo.Context) error {
		ctx := c.Request().Context()
		ep.OrdersMetrics.GetOrdersHttpRequests.Add(ctx, 1)

		listQuery, err := utils.GetListQueryFromCtx(c)
		if err != nil {
			badRequestErr := customErrors.NewBadRequestErrorWrap(
				err,
				"[getOrdersEndpoint_handler.GetListQueryFromCtx] error in getting data from query string",
			)
			ep.Logger.Errorf(
				fmt.Sprintf(
					"[getOrdersEndpoint_handler.GetListQueryFromCtx] err: %v",
					badRequestErr,
				),
			)
			return err
		}

		request := &dtos.GetOrdersRequestDto{ListQuery: listQuery}
		if err := c.Bind(request); err != nil {
			badRequestErr := customErrors.NewBadRequestErrorWrap(
				err,
				"[getOrdersEndpoint_handler.Bind] error in the binding request",
			)
			ep.Logger.Errorf(fmt.Sprintf("[getOrdersEndpoint_handler.Bind] err: %v", badRequestErr))
			return badRequestErr
		}

		query := queries.NewGetOrders(request.ListQuery)

		queryResult, err := mediatr.Send[*queries.GetOrders, *dtos.GetOrdersResponseDto](ctx, query)
		if err != nil {
			err = errors.WithMessage(
				err,
				"[getOrdersEndpoint_handler.Send] error in sending GetOrders",
			)
			ep.Logger.Error(fmt.Sprintf("[getOrdersEndpoint_handler.Send] err: {%v}", err))
			return err
		}

		return c.JSON(http.StatusOK, queryResult)
	}
}